move to separate package @bugsnag/react-router-performance
