+++ packages/platforms/browser/tests/auto-instrumentation/resource-load-plugin.test.ts
@@ -5,7 +5,7 @@
 import { InMemoryDelivery, IncrementingClock, IncrementingIdGenerator, VALID_API_KEY, createTestClient } from '@bugsnag/js-performance-test-utilities'
 import { ResourceLoadPlugin } from '../../lib/auto-instrumentation/resource-load-plugin'
 import { createSchema, type BrowserConfiguration, type BrowserSchema } from '../../lib/config'
-import { createDefaultRoutingProvider } from '../../lib/routing-provider/default-routing-provider'
+import { createDefaultRoutingProvider } from '../../lib/default-routing-provider'
 import createOnSettle from '../../lib/on-settle'
 import { RequestTracker } from '../../lib/request-tracker/request-tracker'
 import { PerformanceFake, PerformanceObserverManager } from '../utilities'
