+++ packages/react-router/rollup.config.npm.mjs
@@ -1,0 +1,5 @@
+import createRollupConfig from '../../.rollup/index.mjs'
+
+export default createRollupConfig({
+    external: ['@bugsnag/browser-performance', 'react-router-dom'],
+})
