+++ test/browser/features/support/build-packages.rb
@@ -24,11 +24,13 @@ $logger.info("Building in #{BUILD_MODE} mode")
 PACKAGE_NAMES = [
   "@bugsnag/core-performance",
   "@bugsnag/browser-performance",
+  "@bugsnag/react-router-performance",
 ]
 
 PACKAGE_DIRECTORIES = [
   "#{ROOT}/packages/core",
   "#{ROOT}/packages/platforms/browser",
+  "#{ROOT}/packages/react-router",
 ]
 
 def run(command)
